+++ releasenotes.md
@@ -2,6 +2,7 @@
 
 #### DOOM Retro v4.6.2
 
+* Minor changes have been made to text that is output to the console.
 * The amount of time the player has been playing the current map is now displayed in the automap if the `am_playerstats` CVAR is `on`.
 
 ![](https://github.com/bradharding/www.doomretro.com/raw/master/wiki/bigdivider.png)